Fix cmake not adding Werror on clang, and _lots_ of warnings (#4963)
* Fix cmake not adding Werror on clang, and _lots_ of warnings

* WIP: Build fixes

* Cannot make intermediate blockhandler instance

* Tiger's changes

* Fix BitIndex check

* Handle invalid NextState values in cMultiVersionProtocol

+++ src/Protocol/Palettes/Palette_1_14.cpp
@@ -9578,7 +9578,7 @@ namespace Palette_1_14
 			case Statistic::WalkOneCm: return 5;
 			case Statistic::WalkOnWaterOneCm: return 8;
 			case Statistic::WalkUnderWaterOneCm: return 12;
-			default: return -1;
+			default: return static_cast<UInt32>(-1);
 		}
 	}
